backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
requests==2.31.0
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.1
